doing setup only once makes tests much faster

+++ src/tests/ggrc/__init__.py
@@ -10,14 +10,22 @@ from ggrc.models import create_db, drop_db
 from ggrc import settings
 
 class TestCase(BaseTestCase):
-  def drop_db(self):
+  done_first_setup = False
+
+  @classmethod
+  def do_first_setup(cls):
+    if cls.done_first_setup:
+      return
+
     db.engine.execute("DROP DATABASE IF EXISTS ggrcdevtest;")
     db.engine.execute("CREATE DATABASE ggrcdevtest; USE ggrcdevtest;")
 
-  def setUp(self):
-    self.drop_db() # if error, tearDown doesn't happen
-
     create_db(use_migrations=True, quiet=True)
+
+    cls.done_first_setup = True
+
+  def setUp(self):
+    TestCase.do_first_setup()
 
     if getattr(settings, 'MEMCACHE_MECHANISM', False) is True:
       from google.appengine.api import memcache
@@ -28,7 +36,6 @@ class TestCase(BaseTestCase):
 
   def tearDown(self):
     db.session.remove()
-    self.drop_db()
 
     if getattr(settings, 'MEMCACHE_MECHANISM', False) is True:
       from google.appengine.api import memcache
